SCons platform support

+++ app/stack.cpp
@@ -19,7 +19,7 @@ void RPNStack::push_items(std::vector<StackItem>&& items) {
 
 
 std::string RPNStack::peek_types(uint64_t count) const {
-    uint64_t safe_count = std::min(count, stack.size());
+    uint64_t safe_count = std::min((size_t)count, stack.size());
 
     std::stringstream ss;
     for (uint64_t idx = safe_count; idx > 0; --idx) {
@@ -35,7 +35,7 @@ std::string RPNStack::peek_types(uint64_t count) const {
 
 
 std::vector<Value::Type> RPNStack::peek_types_vec(uint64_t count) const {
-    uint64_t safe_count = std::min(count, stack.size());
+    uint64_t safe_count = std::min((size_t)count, stack.size());
 
     std::vector<Value::Type> types;
     for (uint64_t idx = safe_count; idx > 0; --idx) {
@@ -50,7 +50,7 @@ size_t RPNStack::size() const { return stack.size(); }
 
 
 std::vector<StackItem> RPNStack::pop_items(uint64_t count) {
-    uint64_t safe_count = std::min(count, stack.size());
+    uint64_t safe_count = std::min((size_t)count, stack.size());
 
     std::vector<StackItem> items;
     for (uint64_t idx = safe_count; idx > 0; --idx) {
